fix(i18n): extract sync helpers out of "use server" actions file
Next.js Turbopack build was failing with "Server Actions must be async
functions" for `parseTranslations`, `buildTranslatableCreate`, and
`buildTranslatableUpdate` — exported sync helpers in a `"use server"`
module.

Move all pure form-parsing / payload-building helpers to a sibling
`helpers.ts` (no "use server" directive). `actions.ts` imports them as
needed. Tests now import from `./helpers` instead of `./actions`.

No behavior change.
