+++ src/lib/generic-catalogs-i18n/actions.test.ts
@@ -3,7 +3,7 @@ import {
   parseTranslations,
   buildTranslatableCreate,
   buildTranslatableUpdate,
-} from "./actions";
+} from "./helpers";
 
 // ─── Helpers ──────────────────────────────────────────────────────────────────
